// frontend/src/app/types/types.ts
export interface User {
    id: string;
    name: string;
    email: string;
}

// Wait for ready
// export interface Product {
//     id: string;
//     name: string;
//     price: number;
// }

// export interface Order {
//     id: string;
//     userId: string;
//     products: Product[];
//     quantity: number,
//     status: string
// }